feat: implement structured error responses with runtime-level validation error interception
This commit implements automatic error classification at the Rust runtime level:
- Application errors (ValidationErrors) return clean, structured JSON at 400
- Runtime errors (bugs) return 500 with error details for debugging

Changes:
- Created rover_types crate for shared ValidationError/ValidationErrors types
- Broke circular dependency between rover_core and rover_server
- Modified guard.rs to throw ValidationErrors as ExternalError (not strings)
- Added runtime error interception in event_loop.rs that:
  - Detects ValidationErrors wrapped in CallbackError
  - Returns structured JSON: {"errors": [{"field": "...", "message": "...", "type": "..."}]}
  - Returns 400 Bad Request for validation errors
  - Returns 500 Internal Server Error for runtime errors
- Simplified Lua code - no changes needed, errors handled automatically

All tests passing (6 tests). Tested with curl on multiple endpoints.
