+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["cli" , "rover_core", "rover_server"]
+members = ["cli" , "rover_core", "rover_server", "rover_types"]
 
 [profile.release]
 opt-level = 3
