+++ rover_core/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2024"
 anyhow = "1.0.100"
 mlua = { version="0.11.5", features = ["anyhow", "vendored", "luajit", "send", "serialize", "async"] }
 rover_server = {path = "../rover_server"}
+rover_types = {path = "../rover_types"}
 serde_json = "1.0"
 
 [lib]
