+++ rover_server/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2024"
 [dependencies]
 anyhow = "1.0.100"
 mlua = { version="0.11.5", features = ["anyhow", "vendored", "luajit", "send", "serialize", "async"] }
+rover_types = { path = "../rover_types" }
 serde_json = "1.0"
 itoa = "1.0"
 ryu = "1.0"
