+++ rover_types/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "rover_types"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+mlua = { version="0.11.5", features = ["anyhow", "vendored", "luajit", "send", "serialize", "async"] }
